Jour 13 - Intégration au projet
Ajout des scripts d'execution pour composer
Mise au niveau 5 de PHPStan et correction des erreurs
Utilisation de pint pour la qualitée
Modification des suggestion par refactor

+++ app/Entity/Product.php
@@ -66,7 +66,7 @@ class Product
         $this->price = $price;
     }
 
-    public function setStock(float $stock): void
+    public function setStock(int $stock): void
     {
         if ($stock < 0) {
             throw new \InvalidArgumentException('Stock négatif interdit');
